Remove duplicate UserInput.cs since methods are in Utils.cs
